make events & clock offsets more robust; feature parity in octave package; update README (PI compat & R2024b compat)

+++ octave/get_ns.m
@@ -1,5 +1,5 @@
 function [ns] = get_ns()
-    ns = double(py.time.time_ns());
+    ns = int64(py.time.time_ns());
 
     return;
 endfunction
